apps/data_quality/requirements.txt
# This file was autogenerated by uv via the following command:
#    uv pip compile requirements.in -o requirements.txt --python-platform=linux
blinker==1.8.2
    # via flask
cachetools==5.5.0
    # via google-auth
certifi==2024.8.30
    # via requests
charset-normalizer==3.3.2
    # via requests
click==8.1.7
    # via
    #   flask
    #   uvicorn
flask==3.0.3
    # via -r requirements.in
google-api-core==1.34.1
    # via
    #   google-api-python-client
    #   google-cloud-appengine-logging
    #   google-cloud-core
    #   google-cloud-logging
    #   google-cloud-secret-manager
    #   google-cloud-storage
google-api-python-client==1.12.8
    # via -r requirements.in
google-auth==2.35.0
    # via
    #   google-api-core
    #   google-api-python-client
    #   google-auth-httplib2
    #   google-cloud-appengine-logging
    #   google-cloud-core
    #   google-cloud-storage
google-auth-httplib2==0.2.0
    # via google-api-python-client
google-cloud-appengine-logging==1.4.5
    # via google-cloud-logging
google-cloud-audit-log==0.3.0
    # via google-cloud-logging
google-cloud-core==2.4.1
    # via
    #   google-cloud-logging
    #   google-cloud-storage
google-cloud-logging==3.5.0
    # via -r requirements.in
google-cloud-secret-manager==2.2.0
    # via -r requirements.in
google-cloud-storage==2.9.0
    # via -r requirements.in
google-crc32c==1.6.0
    # via google-resumable-media
google-resumable-media==2.7.2
    # via google-cloud-storage
googleapis-common-protos==1.65.0
    # via
    #   google-api-core
    #   google-cloud-audit-log
    #   grpc-google-iam-v1
    #   grpcio-status
grpc-google-iam-v1==0.12.7
    # via
    #   google-cloud-logging
    #   google-cloud-secret-manager
grpcio==1.66.1
    # via
    #   google-api-core
    #   googleapis-common-protos
    #   grpc-google-iam-v1
    #   grpcio-status
grpcio-status==1.48.2
    # via google-api-core
gunicorn==20.0.4
    # via -r requirements.in
h11==0.14.0
    # via uvicorn
httplib2==0.22.0
    # via
    #   google-api-python-client
    #   google-auth-httplib2
idna==3.10
    # via requests
itsdangerous==2.2.0
    # via flask
jinja2==3.1.4
    # via flask
libcst==1.4.0
    # via google-cloud-secret-manager
markupsafe==2.1.5
    # via
    #   jinja2
    #   werkzeug
proto-plus==1.24.0
    # via
    #   google-cloud-appengine-logging
    #   google-cloud-logging
    #   google-cloud-secret-manager
protobuf==3.20.3
    # via
    #   google-api-core
    #   google-cloud-appengine-logging
    #   google-cloud-audit-log
    #   google-cloud-logging
    #   googleapis-common-protos
    #   grpc-google-iam-v1
    #   grpcio-status
    #   proto-plus
pyasn1==0.6.1
    # via
    #   pyasn1-modules
    #   rsa
pyasn1-modules==0.4.1
    # via google-auth
pyparsing==3.1.4
    # via httplib2
pyyaml==6.0.2
    # via libcst
requests==2.32.3
    # via
    #   google-api-core
    #   google-cloud-storage
rsa==4.9
    # via google-auth
setuptools==75.1.0
    # via gunicorn
six==1.16.0
    # via google-api-python-client
uritemplate==3.0.1
    # via google-api-python-client
urllib3==2.2.3
    # via requests
uvicorn==0.22.0
    # via -r requirements.in
werkzeug==3.0.4
    # via flask
